% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_estimation.R
\name{compile_cmdstanr_model}
\alias{compile_cmdstanr_model}
\title{Compile Stan Model with cmdstanr}
\usage{
compile_cmdstanr_model(model_path)
}
\arguments{
\item{model_path}{Path to .stan file}
}
\value{
Compiled cmdstanr model
}
\description{
Compile Stan Model with cmdstanr
}
\keyword{internal}
