% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_estimation.R
\name{compile_rstan_model}
\alias{compile_rstan_model}
\title{Compile Stan Model with rstan}
\usage{
compile_rstan_model(model_path)
}
\arguments{
\item{model_path}{Path to .stan file}
}
\value{
Compiled rstan model
}
\description{
Compile Stan Model with rstan
}
\keyword{internal}
