% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgestan_interface.R
\name{create_bridgestan_model}
\alias{create_bridgestan_model}
\title{Create BridgeStan Model Object}
\usage{
create_bridgestan_model(stan_data, model_path = NULL)
}
\arguments{
\item{stan_data}{List of data prepared by \code{\link{prepare_irt_data}}}

\item{model_path}{Optional path to Stan model file. If NULL, uses package model.}
}
\value{
BridgeStan model object
}
\description{
Creates a BridgeStan model for use with external optimizers
}
\details{
BridgeStan provides a lightweight interface to Stan models, allowing use
of external optimization algorithms like optim(), optimx, or nloptr.
}
\examples{
\dontrun{
stan_data <- prepare_irt_data(response_data, loadings, thresh_L, thresh_R)
bs_model <- create_bridgestan_model(stan_data)
}

}
