% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgestan_interface.R
\name{create_stan_json}
\alias{create_stan_json}
\title{Create Stan JSON from Data List}
\usage{
create_stan_json(stan_data)
}
\arguments{
\item{stan_data}{Data list from prepare_irt_data()}
}
\value{
JSON string
}
\description{
Converts R data list to Stan JSON format
}
\keyword{internal}
