% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.get_irtscoring_env}
\alias{.get_irtscoring_env}
\title{Get Package Environment}
\usage{
.get_irtscoring_env()
}
\value{
Environment
}
\description{
Returns the package environment containing compiled Stan models
}
\keyword{internal}
