% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgestan_interface.R
\name{extract_b_from_params}
\alias{extract_b_from_params}
\title{Extract b from Parameter Vector}
\usage{
extract_b_from_params(params, stan_data, bs_model)
}
\description{
Extract b from Parameter Vector
}
\keyword{internal}
