% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgestan_interface.R
\name{extract_beta0_from_params}
\alias{extract_beta0_from_params}
\title{Extract Beta0 from Parameter Vector}
\usage{
extract_beta0_from_params(params, stan_data, bs_model)
}
\description{
Extract Beta0 from Parameter Vector
}
\keyword{internal}
