% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgestan_interface.R
\name{extract_omega_from_params}
\alias{extract_omega_from_params}
\title{Extract Omega from Parameter Vector}
\usage{
extract_omega_from_params(params, stan_data, bs_model)
}
\description{
Extract Omega from Parameter Vector
}
\keyword{internal}
