% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgestan_interface.R
\name{extract_tau_from_params}
\alias{extract_tau_from_params}
\title{Extract Tau from Parameter Vector}
\usage{
extract_tau_from_params(params, stan_data, bs_model)
}
\description{
Extract Tau from Parameter Vector
}
\keyword{internal}
