% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgestan_interface.R
\name{extract_theta_from_params}
\alias{extract_theta_from_params}
\title{Extract Theta from Parameter Vector}
\usage{
extract_theta_from_params(params, stan_data, bs_model)
}
\description{
Extract Theta from Parameter Vector
}
\keyword{internal}
