% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_estimation.R
\name{extract_vb_samples}
\alias{extract_vb_samples}
\title{Extract VB Samples}
\usage{
extract_vb_samples(fit, backend)
}
\arguments{
\item{fit}{VB fit object}

\item{backend}{Backend used}
}
\value{
List of parameter samples
}
\description{
Helper to extract samples from VB fit objects
}
\keyword{internal}
