% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_estimation.R
\name{fit_irt_laplace}
\alias{fit_irt_laplace}
\title{Fit IRT Model using Laplace Approximation}
\usage{
fit_irt_laplace(stan_data, backend = NULL, draws = 1000, ...)
}
\arguments{
\item{stan_data}{List of data prepared by \code{\link{prepare_irt_data}}}

\item{backend}{Stan backend to use: "cmdstanr" or "rstan"}

\item{draws}{Number of draws to sample from the Laplace approximation}

\item{...}{Additional arguments passed to the optimizer}
}
\value{
Object of class "irt_fit" containing MAP estimates plus:
\item{posterior_samples}{Matrix of draws from Laplace approximation}
\item{hessian}{Hessian matrix at MAP estimate}
}
\description{
Fits multidimensional IRT model using Laplace approximation. Finds the MAP
estimate then approximates the posterior with a multivariate normal.
}
\examples{
\dontrun{
fit <- fit_irt_laplace(stan_data, draws = 1000)
}

}
