% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_estimation.R
\name{fit_irt_map}
\alias{fit_irt_map}
\title{Fit IRT Model using MAP Estimation}
\usage{
fit_irt_map(stan_data, backend = NULL, ...)
}
\arguments{
\item{stan_data}{List of data prepared by \code{\link{prepare_irt_data}}}

\item{backend}{Stan backend to use: "cmdstanr" or "rstan"}

\item{...}{Additional arguments passed to the optimizer}
}
\value{
Object of class "irt_fit" containing:
\item{theta}{N x K matrix of estimated ability scores}
\item{beta0}{Length K vector of dimension intercepts}
\item{b}{K x J matrix of covariate effects}
\item{tau}{Length K vector of dimension standard deviations}
\item{Omega}{K x K correlation matrix}
\item{log_lik}{Log-likelihood at MAP estimate}
}
\description{
Fits multidimensional IRT model using Maximum A Posteriori (MAP) estimation.
This is the fastest method, finding the mode of the posterior distribution.
}
\examples{
\dontrun{
# Prepare data
stan_data <- prepare_irt_data(response_data, loadings, thresh_L, thresh_R)

# Fit using MAP
fit <- fit_irt_map(stan_data)

# Extract ability estimates
theta_hat <- fit$theta
}

}
