% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_estimation.R
\name{fit_irt_mcmc}
\alias{fit_irt_mcmc}
\title{Fit IRT Model using MCMC}
\usage{
fit_irt_mcmc(
  stan_data,
  backend = NULL,
  chains = 4,
  iter = 2000,
  warmup = 1000,
  ...
)
}
\arguments{
\item{stan_data}{List of data prepared by \code{\link{prepare_irt_data}}}

\item{backend}{Stan backend to use: "cmdstanr" or "rstan"}

\item{chains}{Number of MCMC chains to run}

\item{iter}{Number of iterations per chain (rstan) or total (cmdstanr)}

\item{warmup}{Number of warmup iterations}

\item{...}{Additional arguments passed to MCMC sampler}
}
\value{
Object of class "irt_fit" containing MCMC samples
}
\description{
Fits multidimensional IRT model using full Bayesian MCMC sampling.
This is the most computationally intensive but provides complete
posterior distributions.
}
\examples{
\dontrun{
# Run MCMC with 4 chains
fit <- fit_irt_mcmc(stan_data, chains = 4, iter = 2000)
}

}
