% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgestan_interface.R
\name{fit_irt_nloptr}
\alias{fit_irt_nloptr}
\title{Fit IRT Model Using nloptr via BridgeStan}
\usage{
fit_irt_nloptr(
  stan_data,
  algorithm = "NLOPT_LD_LBFGS",
  init = NULL,
  opts = list(),
  ...
)
}
\arguments{
\item{stan_data}{List of data prepared by \code{\link{prepare_irt_data}}}

\item{algorithm}{Algorithm for nloptr (default: "NLOPT_LD_LBFGS")}

\item{init}{Optional initial parameter values}

\item{opts}{List of options passed to nloptr}

\item{...}{Additional arguments}
}
\value{
Object of class "irt_fit"
}
\description{
Uses nloptr package algorithms with BridgeStan interface
}
\details{
Provides access to advanced optimization algorithms from the NLopt library
via nloptr package. Common algorithms:
\itemize{
\item "NLOPT_LD_LBFGS": Low-storage BFGS
\item "NLOPT_LD_VAR1": Shifted limited-memory variable-metric
\item "NLOPT_LD_TNEWTON_PRECOND": Truncated Newton with preconditioning
}

Requires both bridgestan and nloptr packages.
}
\examples{
\dontrun{
fit <- fit_irt_nloptr(stan_data, algorithm = "NLOPT_LD_LBFGS")
}

}
