% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgestan_interface.R
\name{fit_irt_optim}
\alias{fit_irt_optim}
\title{Fit IRT Model Using optim() via BridgeStan}
\usage{
fit_irt_optim(
  stan_data,
  method = "L-BFGS-B",
  init = NULL,
  control = list(),
  ...
)
}
\arguments{
\item{stan_data}{List of data prepared by \code{\link{prepare_irt_data}}}

\item{method}{Optimization method for optim() (default: "L-BFGS-B")}

\item{init}{Optional initial parameter values}

\item{control}{List of control parameters passed to optim()}

\item{...}{Additional arguments passed to optim()}
}
\value{
Object of class "irt_fit"
}
\description{
Uses base R optim() function with BridgeStan interface for optimization
}
\details{
This function provides access to R's built-in optimization algorithms
via the BridgeStan interface. Common methods include:
\itemize{
\item "L-BFGS-B": Limited-memory BFGS with box constraints
\item "BFGS": Quasi-Newton method
\item "CG": Conjugate gradient
}

Requires the bridgestan package to be installed.
}
\examples{
\dontrun{
stan_data <- prepare_irt_data(response_data, loadings, thresh_L, thresh_R)
fit <- fit_irt_optim(stan_data, method = "L-BFGS-B")
}

}
