% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_estimation.R
\name{fit_irt_vb}
\alias{fit_irt_vb}
\title{Fit IRT Model using Variational Bayes}
\usage{
fit_irt_vb(
  stan_data,
  backend = NULL,
  algorithm = c("pathfinder", "meanfield"),
  draws = 1000,
  ...
)
}
\arguments{
\item{stan_data}{List of data prepared by \code{\link{prepare_irt_data}}}

\item{backend}{Stan backend to use: "cmdstanr" or "rstan"}

\item{algorithm}{VB algorithm: "pathfinder" (cmdstanr only) or "meanfield"}

\item{draws}{Number of draws to sample from variational distribution}

\item{...}{Additional arguments passed to VB algorithm}
}
\value{
Object of class "irt_fit" containing variational estimates
}
\description{
Fits multidimensional IRT model using Variational Bayes (VB) approximation.
Tries to use Pathfinder algorithm if available, otherwise falls back to ADVI.
}
\examples{
\dontrun{
# Use Pathfinder (fast, cmdstanr only)
fit <- fit_irt_vb(stan_data, algorithm = "pathfinder")

# Use mean-field ADVI
fit <- fit_irt_vb(stan_data, algorithm = "meanfield")
}

}
