% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_interface.R
\name{fit_unidim_ability}
\alias{fit_unidim_ability}
\title{Fit Unidimensional IRT Model}
\usage{
fit_unidim_ability(
  response_data,
  item_discrimination,
  threshold_left,
  threshold_right,
  ...
)
}
\arguments{
\item{response_data}{Data frame with columns \code{pid} and \code{iid}}

\item{item_discrimination}{Numeric vector of discrimination parameters (length L)}

\item{threshold_left}{Numeric vector of left thresholds (length L)}

\item{threshold_right}{Numeric vector of right thresholds (length L)}

\item{...}{Additional arguments passed to \code{\link{fit_ability}}}
}
\value{
Object of class "irt_fit"
}
\description{
Convenience wrapper for unidimensional IRT models. Accepts discrimination
parameters as a vector and converts to the L x 1 loading matrix format.
}
\examples{
\dontrun{
response_data <- data.frame(
  pid = c(1, 1, 2, 2),
  iid = c(1, 2, 1, 2)
)

fit <- fit_unidim_ability(
  response_data = response_data,
  item_discrimination = c(1, 1.5, 1, 1.5),
  threshold_left = c(-Inf, -1, -Inf, -1),
  threshold_right = c(-1, Inf, -1, Inf),
  method = "map"
)
}

}
