% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_covariate_effects}
\alias{get_covariate_effects}
\title{Get Covariate Effects from IRT Fit}
\usage{
get_covariate_effects(fit, stan_data = NULL)
}
\arguments{
\item{fit}{An object of class "irt_fit"}

\item{stan_data}{Optional stan_data list. If NULL, attempts to reconstruct
from fit object.}
}
\value{
N x K matrix of covariate effects
}
\description{
Calculate the contribution of covariates to ability scores (beta0 + X*b)
}
\keyword{internal}
