% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_dimension_correlations}
\alias{get_dimension_correlations}
\title{Get Dimension Correlations from IRT Fit}
\usage{
get_dimension_correlations(fit)
}
\arguments{
\item{fit}{An object of class "irt_fit"}
}
\value{
K x K correlation matrix
}
\description{
Extract the correlation matrix between dimensions (Omega)
}
\keyword{internal}
