% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_individual_effects}
\alias{get_individual_effects}
\title{Get Individual Effects from IRT Fit}
\usage{
get_individual_effects(fit)
}
\arguments{
\item{fit}{An object of class "irt_fit"}
}
\value{
N x K matrix of individual effects
}
\description{
Extract the individual-specific deviations (zeta/eta component)
}
\keyword{internal}
