% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_estimation.R
\name{get_stan_backend}
\alias{get_stan_backend}
\title{Get Current Stan Backend}
\usage{
get_stan_backend()
}
\value{
Character string: "cmdstanr", "rstan", or NULL if no model compiled
}
\description{
Get Current Stan Backend
}
