% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgestan_interface.R
\name{has_bridgestan}
\alias{has_bridgestan}
\title{Check if BridgeStan is Available}
\usage{
has_bridgestan()
}
\value{
Logical indicating if bridgestan package is installed
}
\description{
Check if BridgeStan is Available
}
