% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_estimation.R
\name{has_cmdstanr}
\alias{has_cmdstanr}
\title{Check if cmdstanr is Available}
\usage{
has_cmdstanr()
}
\value{
Logical indicating if cmdstanr package is installed
}
\description{
Check if cmdstanr is Available
}
