% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_estimation.R
\name{has_pathfinder}
\alias{has_pathfinder}
\title{Check if Pathfinder is Available}
\usage{
has_pathfinder()
}
\value{
Logical indicating if cmdstanr has Pathfinder support
}
\description{
Check if Pathfinder is Available
}
\keyword{internal}
