% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_estimation.R
\name{is_stan_model_compiled}
\alias{is_stan_model_compiled}
\title{Check if Stan Model is Compiled}
\usage{
is_stan_model_compiled()
}
\value{
Logical indicating whether the Stan model has been compiled
}
\description{
Check if Stan Model is Compiled
}
