% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_estimation.R
\name{new_irt_fit}
\alias{new_irt_fit}
\title{IRT Model Fit Object}
\usage{
new_irt_fit(
  theta,
  beta0,
  b,
  tau,
  Omega,
  log_lik,
  method,
  backend,
  stan_fit = NULL,
  ...
)
}
\arguments{
\item{theta}{Matrix of estimated ability parameters (N x K)}

\item{beta0}{Vector of dimension intercepts (length K)}

\item{b}{Matrix of covariate effects (K x J)}

\item{tau}{Vector of dimension standard deviations (length K)}

\item{Omega}{Correlation matrix of dimensions (K x K)}

\item{log_lik}{Log-likelihood value}

\item{method}{Estimation method used}

\item{backend}{Stan backend used}

\item{stan_fit}{Original Stan fit object (optional)}

\item{...}{Additional components}
}
\value{
Object of class "irt_fit"
}
\description{
S3 class for storing IRT model fit results
}
\keyword{internal}
