% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_estimation.R
\name{parse_parameter_samples}
\alias{parse_parameter_samples}
\title{Parse Parameter Samples from Optimization Fit}
\usage{
parse_parameter_samples(fit, draws)
}
\arguments{
\item{fit}{Stan optimization fit object}

\item{draws}{Number of draws to sample}
}
\value{
Matrix of parameter samples
}
\description{
Helper function to extract parameter samples for Laplace approximation
}
\keyword{internal}
