% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.irt_fit}
\alias{print.irt_fit}
\title{Print IRT Model Fit}
\usage{
\method{print}{irt_fit}(x, ...)
}
\arguments{
\item{x}{An object of class "irt_fit"}

\item{...}{Additional arguments (currently unused)}
}
\value{
Invisibly returns the input object
}
\description{
Print IRT Model Fit
}
