% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_estimation.R
\name{run_pathfinder}
\alias{run_pathfinder}
\title{Run Pathfinder Algorithm}
\usage{
run_pathfinder(model, stan_data, draws, ...)
}
\arguments{
\item{model}{Compiled cmdstanr model}

\item{stan_data}{Data list}

\item{draws}{Number of draws}

\item{...}{Additional arguments}
}
\value{
Pathfinder fit object
}
\description{
Run Pathfinder Algorithm
}
\keyword{internal}
