% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.irt_fit}
\alias{summary.irt_fit}
\title{Summary of IRT Model Fit}
\usage{
\method{summary}{irt_fit}(object, ...)
}
\arguments{
\item{object}{An object of class "irt_fit"}

\item{...}{Additional arguments (currently unused)}
}
\value{
Invisibly returns the input object
}
\description{
Summary of IRT Model Fit
}
