% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{validate_item_loadings}
\alias{validate_item_loadings}
\title{Validate Item Loadings}
\usage{
validate_item_loadings(loadings, K)
}
\arguments{
\item{loadings}{Loading matrix to validate}

\item{K}{Number of dimensions}
}
\description{
Checks that item loadings matrix has the correct structure
}
\keyword{internal}
