% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{validate_response_data}
\alias{validate_response_data}
\title{Validate Response Data}
\usage{
validate_response_data(data)
}
\arguments{
\item{data}{Data frame to validate}
}
\description{
Checks that response data has the required structure
}
\keyword{internal}
