% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{validate_thresholds}
\alias{validate_thresholds}
\title{Validate Thresholds}
\usage{
validate_thresholds(left, right, L)
}
\arguments{
\item{left}{Left thresholds}

\item{right}{Right thresholds}

\item{L}{Number of responses}
}
\description{
Checks that threshold vectors have correct structure and ordering
}
\keyword{internal}
